Optimize looping logic for GUI dashboard location filters
- This will connect all filter types at once
- We're reducing the number of checks for the default filter value to just one
