+++ frontend/test/metabase/scenarios/dashboard-filters/helpers/e2e-dashboard-filter-data-objects.js
@@ -35,30 +35,3 @@ export const DASHBOARD_DATE_FILTERS = {
     representativeResult: "37.65",
   },
 };
-
-export const DASHBOARD_LOCATION_FILTERS = {
-  Dropdown: {
-    value: "Abbeville",
-    representativeResult: "1510",
-  },
-  "Is not": {
-    value: "Abbeville",
-    representativeResult: "37.65",
-  },
-  Contains: {
-    value: "Abb",
-    representativeResult: "1510",
-  },
-  "Does not contain": {
-    value: "Abb",
-    representativeResult: "37.65",
-  },
-  "Starts with": {
-    value: "Abb",
-    representativeResult: "1510",
-  },
-  "Ends with": {
-    value: "y",
-    representativeResult: "115.24",
-  },
-};
